App structure, basic engine code.
Solstice Submarine CTF is a game that I am making for the Github Game
Off 2012, and that will also be a part of the third installment of
Solstid Submarine digital interactive content that I have been creating
annually in partnership with the incomparable Donna Almendrala.

The game as planned is a multi-player simul-turn-based strategy game
reminiscent of traditional capture the flag games.