Scene rendering pipeline. Entity cleanup.
Three.js has been mostly abandoned as the base rendering framework. It
turns out that a 3D rendering framework just isn't the best thing for
getting that retro 2D look :)

A custom scene graph rendering pipeline has been put in place as a
stand-in for what Three.js was supposed to be doing. So far it looks
like it will work out just fine.