Full map. New render pipeline. Entities.
I have tweaked the rendering pipeline to move it in the direction of
being able to update only changed regions of the viewport. This will
probably result in a memory usage trade off for the sake of performance.
Particularly, I'm seeking to have high visual performance on mobile
devices. It would be awesome of this game ran at 60 FPS in a phone's
browser (and Mobile FF on Android is close already).

The map is now close to where it will probably be when the game is
shipped. Also, stubs for important game entites have been added. Wall
graphics have been flattened into the sand for the time being until a
performance profile suggests that this can be undone. The Env class has
been renamed to what it actually is: Renderer.