Animated sprites. Selective scene re-rendering.
Animated sprites have been added. The only animated sprite visible right
now is the randomly generated grass / seaweed, but there are animation
frames for the characters as well if you look at the sprites.

In an effort to improve performance, scene re-rendering has been
switched to an event model where child nodes of the scene inform the
renderer when a redraw of an area of the scene needs to happen. I am not
sure that I am doing this in the most efficient way possible, but
animations are enabled and we are getting as high an FPS as ever on
mobile.

Rotation is broken as of this commit. Hopefully I will have time to fix
it this week.