Improved drawing ratios. Removed button handling.
The calculations that determined drawing ratios were incorrect for 1:1
pixel density screens. This has been fixed. JavaScript hooks into media
queries have been added to allow for dynamic resizing of the renderer as
the screen resizes. The end result of these things is that on an iPhone
3GS, I am getting 30+ FPS and drawing at 1:1 pixel density on a small
canvas, while on more powerful / larger-screened devices I am scaling up
pixel density and canvas dimensions proportionally.

Button handling was removed, as the single method of input for the game
as it stands will be the mouse.