Render pipeline fix. Click fix.
The render pipeline had a pretty severe bug. Redraw rectangles were
being substantially inflated beyond the correct size. Also, some tiles
were not properly redrawing when appropriate. A debugging mode has been
added to the render pipeline to show what areas are redrawn when redraw
happens. With this fix in, I am getting in excess of 60 FPS on mobile
Firefox, and roughly 30 FPS on mobile Chrome.

Clicking was only working in Chrome on account of me using the wrong
properties to stash the x and y coordinates relative to the element.
StackOverflow saved the day and now clicking works in most browsers and
devices that I have tried.