Hero movement. Rendering bugfix.
The hero units can now move along a path to a selected destination.
Currently the pathing algorithm is kinda broken, but it is a good
demonstration of the desired look and feel, and also makes for a useful
benchmarking milestone.

Rendering has been adjusted to use the canvas context's save and restore
methods, which it turns out I was accidentally recreating as part of my
render pipeline. I realised this was a problemwhen I noticed rounding
errors start to aggregate in my measurement of total aggregate rotation
while drawing. This resulted in the scene moving off center over time.