Waypoint integration. Fog of war stubs.
Waypoint functionality has been more or less completed. Waypoints show
where a player intends to move, and are sensitive to the distinction
between shielded and unshielded movement. Stubs for fog of war have been
added, in preparation for what will likely be the largest
non-multiplayer feature left to implement in the game.