Basic AI. Readme additions. Rule enforcement.
We now have a basic AI. It moves the rockets back and forth between two
different tiles. Not much, but it is a start. The basic AI is a
placeholder to facilitate development of the Arbiter class, which is the
stateless singleton that receives the game state and churns out the
outcomes that game clients should display. Finally, local installation
instructions have been added to the readme.