Arcade panels. Dispose bugfix. Fork work.
Arcade panels have been integrated. Thanks to Donna Almendrala for
drawing such awesome pieces for the game! A bug that was causing
out-of-order disposal has been fixed. More work has been done to get
fork mechanics in place, but they are still a way out. We are gonna be
cutting it close!