Forks. Sprites. Points.
Basic fork interaction is in place. Forks can be picked up, dropped and
captured. Additionally, points are now allocated when you capture a
fork, when you kill an enemy and when you block an attack with your
shield. Similarly, points are subtracted when you die. Artistic notes
have been added to the readme, as well as an embedded rendering of all
of the sprites used in the game.